Added a custom new parser backend. The rationale was that the current parser has a high memory peak when the DataFrames are added to a single chunk. The peak memory also lead to a bloated heap. This new implementation is less memory hungry and is faster. It is not dependent on exernal libraries. It is also not CSV complient. It still needs to implement proper escaping of quotes for instance.
