+++ polars/polars-io/src/csv_core/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod buffer;
+mod chunked_parser;
+pub mod csv;
+pub(crate) mod parser;
+pub mod utils;
