+++ polars/polars-io/src/csv_core/parser.rs
@@ -1,0 +1,271 @@
+use super::buffer::*;
+use num::traits::Pow;
+use polars_core::prelude::*;
+
+/// Skip the utf-8 Byte Order Mark.
+/// credits to csv-core
+pub(crate) fn skip_bom(input: &[u8]) -> &[u8] {
+    if input.len() >= 3 && &input[0..3] == b"\xef\xbb\xbf" {
+        &input[..3]
+    } else {
+        input
+    }
+}
+
+pub(crate) fn all_digit(string: &str) -> bool {
+    string.chars().all(|c| c.is_ascii_digit())
+}
+
+pub(crate) fn next_line_position(input: &[u8]) -> Option<usize> {
+    let pos = input.iter().position(|b| *b == b'\n')?;
+    input.get(pos + 1).and_then(|&b| {
+        Option::from({
+            if b == b'\r' {
+                pos + 1
+            } else {
+                pos
+            }
+        })
+    })
+}
+
+pub(crate) fn is_line_ending(b: u8) -> bool {
+    b == b'\n' || b == b'\r'
+}
+
+pub(crate) fn is_whitespace(b: u8) -> bool {
+    b == b' ' || b == b'\t'
+}
+
+fn skip_condition<F>(input: &[u8], f: F) -> (&[u8], usize)
+where
+    F: Fn(u8) -> bool,
+{
+    let mut read = 0;
+    loop {
+        let b = input[read];
+        if !f(b) {
+            break;
+        }
+        read += 1;
+    }
+    (&input[read..], read)
+}
+
+/// Makes sure that the bytes stream starts with
+///     'field_1,field_2'
+/// and not with
+///     '\nfield_1,field_1'
+pub(crate) fn skip_header(input: &[u8]) -> (&[u8], usize) {
+    let mut pos = next_line_position(input).expect("no lines in the file");
+    if input[pos + 1] == b'\n' {
+        pos += 1;
+    }
+    (&input[pos + 1..], pos + 1)
+}
+
+/// Remove whitespace and line endings from the start of file.
+pub(crate) fn skip_whitespace(input: &[u8]) -> (&[u8], usize) {
+    skip_condition(input, |b| is_whitespace(b) || is_line_ending(b))
+}
+
+pub(crate) fn skip_line_ending(input: &[u8]) -> (&[u8], usize) {
+    skip_condition(input, is_line_ending)
+}
+
+/// Get the mean and standard deviation of length of lines in bytes
+pub(crate) fn get_line_stats(mut bytes: &[u8], n_lines: usize) -> Option<(f32, f32)> {
+    let mut n_read = 0;
+    let mut lengths = Vec::with_capacity(n_lines);
+
+    for _ in 0..n_lines {
+        if n_read >= bytes.len() {
+            return None;
+        }
+        bytes = &bytes[n_read..];
+        match bytes.iter().position(|&b| b == b'\n') {
+            Some(position) => {
+                n_read += position + 1;
+                lengths.push(position + 1);
+            }
+            None => {
+                return None;
+            }
+        }
+    }
+    let mean = (n_read as f32) / (n_lines as f32);
+    let mut std = 0.0;
+    for &len in lengths.iter().take(n_lines) {
+        std += (len as f32 - mean).pow(2.0)
+    }
+    std = (std / n_lines as f32).pow(0.5);
+    Some((mean, std))
+}
+
+pub(crate) fn parse_lines(
+    bytes: &[u8],
+    offset: usize,
+    delimiter: u8,
+    projection: &[usize],
+    buffers: &mut [Buffer],
+    ignore_parser_errors: bool,
+) -> Result<()> {
+    // This variable will store the number of bytes we read. It is important to do this bookkeeping
+    // to be able to correctly parse the strings later.
+    let mut read = offset;
+
+    // We split the lines by the new line characters in the outer loop.
+    // in the inner loop we deal with the fields/columns.
+    // Any primitive type is directly parsed and stored in a Vec buffer.
+    // String types are not parsed. We store strings the starting index in the bytes array and store
+    // the length of the string field. We also store the total length of processed string fields per column.
+    // Later we use that meta information to exactly allocate the required buffers and parse the strings.
+    for mut line in bytes.split(|b| *b == b'\n') {
+        let len = line.len();
+
+        // two adjacent '\n\n' will lead to an empty line.
+        if len == 0 {
+            read += 1;
+            continue;
+        }
+        // including the '\n' character
+        let line_length = len + 1;
+
+        let trailing_byte = line[len - 1];
+        if trailing_byte == b'\r' {
+            line = &line[..len - 1];
+        }
+
+        // read at start of the line
+        let read_sol = read;
+        // // +1 is the split character
+        // read += 1;
+        let leading_byte = line[0];
+
+        // Check if there are some leading character we don't expect on the start of the line.
+        //
+        // # Expected lines
+        //      'field_0,field_1\n'
+        //      'field_0,field_1\r\n'
+        //
+        // # Unexpected lines
+        //      leading delimiter
+        //      ',field_0,field_1'
+        //
+        if leading_byte == delimiter {
+            read += 1;
+            line = &line[1..]
+        }
+
+        // Every line we only need to parse the columns that are projected.
+        // Therefore we check if the idx of the field is in our projected columns.
+        // If it is not, we skip the field.
+        let mut projection_iter = projection.iter().copied();
+        let mut next_projected = projection_iter
+            .next()
+            .expect("at least one column should be projected");
+        let mut processed_fields = 0;
+
+        for (idx, field) in line.split(|b| *b == delimiter).enumerate() {
+            if idx == next_projected {
+                // todo! get rid of bound check
+                let buf = &mut buffers[processed_fields];
+                buf.add(field, ignore_parser_errors, read)?;
+
+                processed_fields += 1;
+
+                // if we have all projected columns we are done with this line
+                match projection_iter.next() {
+                    Some(p) => next_projected = p,
+                    None => {
+                        break;
+                    }
+                }
+            }
+            // +1 is the split character that is consumed by the iterator.
+            read += field.len() + 1;
+        }
+        // this way we also include the trailing '\n' or '\r\n' in the bytes read
+        // and any skipped fields.
+        read = read_sol + line_length;
+    }
+    Ok(())
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+    use std::io::Read;
+
+    #[test]
+    fn test_skip() {
+        let input = b"    hello";
+        assert_eq!(skip_whitespace(input).0, b"hello");
+        let input = b"\n        hello";
+        assert_eq!(skip_whitespace(input).0, b"hello");
+        let input = b"\t\n\r
+        hello";
+        assert_eq!(skip_whitespace(input).0, b"hello");
+    }
+
+    #[test]
+    fn test_parse_lines() {
+        let path = "../../examples/aggregate_multiple_files_in_chunks/datasets/foods1.csv";
+        let mut file = std::fs::File::open(path).unwrap();
+        let mut input = vec![];
+        file.read_to_end(&mut input).unwrap();
+        let bytes = skip_header(skip_whitespace(&input).0).0;
+
+        // after a skip header, we should not have a new line char.
+        assert_ne!(bytes[0], b'\n');
+        dbg!(std::str::from_utf8(bytes));
+
+        let mut buffers = vec![
+            Buffer::Utf8(vec![], 0),
+            Buffer::UInt64(vec![]),
+            Buffer::Float64(vec![]),
+            Buffer::Float64(vec![]),
+        ];
+
+        macro_rules! call_buff_method {
+            ($buf: expr, $method:ident) => {{
+                use Buffer::*;
+                match $buf {
+                    Boolean(a) => a.$method(),
+                    Int32(a) => a.$method(),
+                    Int64(a) => a.$method(),
+                    UInt64(a) => a.$method(),
+                    UInt32(a) => a.$method(),
+                    Float32(a) => a.$method(),
+                    Float64(a) => a.$method(),
+                    Utf8(a, _) => a.$method(),
+                }
+            }};
+        }
+
+        let projection = &[0, 1, 2, 3];
+
+        parse_lines(&bytes, 0, b',', projection, &mut buffers, false).unwrap();
+        // check if all buffers are correctly filled.
+        for buf in &buffers {
+            let len = call_buff_method!(buf, len);
+            assert_eq!(len, 27);
+        }
+
+        dbg!((&buffers[0]));
+        // check if we can reconstruct the correct strings from the accumulated offsets.
+        if let Buffer::Utf8(buf, len) = &buffers[0] {
+            let v = buf
+                .iter()
+                .map(|utf8_field| utf8_field.parse_str(bytes).unwrap())
+                .collect::<Vec<_>>();
+
+            let total_len: usize = v.iter().map(|s| s.len()).sum();
+            assert_eq!(total_len, *len);
+
+            assert_eq!(&v[0], &"vegetables");
+            assert_eq!(&v[v.len() - 1], &"fruit");
+            dbg!(v);
+        }
+    }
+}
