--- polars/polars-io/src/fork/mod.rs
@@ -1,1 +1,0 @@
-pub mod csv;
